Rename executable-schema -> test-schema, used in integration tests
