+++ test/integration/test-helpers/test-schema.ts
@@ -3,5 +3,5 @@ import { readFileSync } from 'fs';
 import path from 'path';
 
 const schemaPath = path.resolve(__dirname, 'schema.graphql');
-export const typeDefs = readFileSync(schemaPath, 'utf-8');
+const typeDefs = readFileSync(schemaPath, 'utf-8');
 export const graphqlSchema = buildSchema(typeDefs);
